+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and the project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+ - Add a way to globally configure transitions between pages while in a PreferenceScreen.
 
 ## [0.4.0] - 2020-11-10
 ### Added